move TLM protocol definition into types directory
